Add optional constructor argument to BinaryEncoder to set dataCodec
